usb driver: Only board specific function of cable config is in SYS.
